Add prop to clear TextInput on focus
If we're using complex objects, editing the box might clear the object, meaning a new object needs to be selected. In which case clearing the input to indicate something needs selecting again might be an appropriate solution.

Adds the prop `clearOnFocus` to `AutoSuggest` which is a boolean, this also allows the possibility of programmatically deciding whether it should be cleared or not, such as in cases where an object has yet to be selected.